Changed double to long double

+++ includes/ft_printf_types.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf_types.h                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dda-silv <dda-silv@student.42.fr>          +#+  +:+       +#+        */
+/*   By: dda-silv <dda-silv@student.42lisboa.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/23 17:17:14 by dda-silv          #+#    #+#             */
-/*   Updated: 2021/01/25 08:05:20 by dda-silv         ###   ########.fr       */
+/*   Updated: 2021/02/18 17:15:34 by dda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,8 @@ void	print_pct(t_format *settings);
 void	print_flt(t_format *settings);
 void	print_exp(t_format *settings);
 void	print_g(t_format *settings);
-void	print_g_flt(t_format *settings, double nb, int precision);
-void	print_g_exp(t_format *settings, double nb, int precision);
+void	print_g_flt(t_format *settings, long double nb, int precision);
+void	print_g_exp(t_format *settings, long double nb, int precision);
 void	print_left_right(t_format *settings, char *str);
 
 #endif
